refactor(plugin): extract inner DTOs from FlipSmartApiClient into api/dto
Part of #730 (epic #728).

- Extract the 12 public nested DTOs (AuthResult, DeviceAuthResponse,
  DeviceStatusResponse, MotdChannelData, MotdResponse, TransactionRequest
  (+Builder), HistoryBackfillEntry, DumpsResponse, BankItem, BankItemId,
  WikiPrice, FlipAdjustmentRequest) into top-level files in
  com.flipsmart.api.dto; update all FlipSmartApiClient.<Name> references.
- Only the two private cache helpers (CachedAnalysis, CachedDailyVolume)
  remain nested.
- Device DTO access goes through existing getters/setters; request DTO
  fields widened to public final to preserve direct-read access. No
  behavior change.
