+++ src/main/java/com/flipsmart/ActiveOfferAdvisorService.java
@@ -1,4 +1,5 @@
 package com.flipsmart;
+import com.flipsmart.api.dto.WikiPrice;
 import com.flipsmart.domain.offer.TrackedOffer;
 import com.flipsmart.api.dto.OfferAdviceResponse;
 import com.flipsmart.domain.offer.OfferDisposition;
@@ -82,7 +83,7 @@ public class ActiveOfferAdvisorService
 
 	static OfferAdviceRequest buildSnapshot(
 		TrackedOffer offer,
-		FlipSmartApiClient.WikiPrice market,
+		WikiPrice market,
 		Integer userAvgBuyPrice,
 		Integer dailyVolume)
 	{
